+++ test/cases/msg-exception-loader-beforePreprocessor/webpack.config.js
@@ -1,0 +1,17 @@
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/index.html',
+      },
+      loaderOptions: {
+        // test error in beforePreprocessor
+        beforePreprocessor: (content, { data }) => UndefinedFuncion(),
+      },
+    }),
+  ],
+};